Refactor Hibernate metrics config and fix opportunity response type

+++ src/main/java/com/candido/server/domain/v1/_common/Ownable.java
@@ -1,5 +1,5 @@
 package com.candido.server.domain.v1._common;
 
 public interface Ownable {
-    Integer getOwnerId();
+    Long getOwnerId();
 }
