+++ src/main/java/com/candido/server/service/base/auth/provider/AuthProviderService.java
@@ -1,5 +1,5 @@
 package com.candido.server.service.base.auth.provider;
 
 public interface AuthProviderService {
-    void addProviderToAccount(int providerId, int accountId);
+    void addProviderToAccount(int providerId, Long accountId);
 }
